Update weather example to read from API or from stored JSON results.
There was an issue with calling the API too many times in one day. Then, the result from the API call would be none.
To solve this, I save the weather to a json file and only update it when the time has changed by an amount that will ensure we never hit the API limit.

Additionally, just in case there is another bug somewhere, I check that the returned value is not None before trying to print it.